services/api/pyproject.toml
[project]
name = "listmonk"
version = "0.3.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = ["fastapi[standard]>=0.128.0", "pydantic-settings>=2.12.0", "taskipy>=1.14.1", "ruff>=0.14.14", "pytest-cov>=7.0.0", "pytest>=9.0.2", "pytest-asyncio>=1.3.0", "factory-boy>=3.3.3", "freezegun>=1.5.5", "httpx>=0.28.1", "greenlet>=3.3.1", "requests>=2.32.5", "pocketbase>=0.15.0", "opentelemetry-sdk>=1.40.0", "opentelemetry-exporter-otlp-proto-grpc>=1.40.0", "opentelemetry-instrumentation-fastapi>=0.61b0", "opentelemetry-instrumentation-logging>=0.61b0", "opentelemetry-instrumentation-requests>=0.61b0", "scalar-fastapi>=1.8.1"]
requires-python = ">=3.14"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
[tool.ruff]
line-length = 128
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.lint.per-file-ignores]
# OTel SDK exposes its logging API via private modules (_logs, _log_exporter) at
# the installed SDK version — no public opentelemetry.sdk.logs module exists yet.
'app/telemetry.py' = ['PLC2701']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/main.py'
pre_test = 'task format'
test = 'pytest -s -x --cov=app --cov-report=xml -vv'
post_test = 'coverage html'

[tool.pdm]
distribution = true
